requirements.txt
fastapi
uvicorn[standard]
jinja2
pyrefly
python-multipart
google-genai
python-dotenv
slowapi
